+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/affixes/adders/JewelrySuffixes.java
@@ -4,10 +4,7 @@ import com.robertx22.age_of_exile.aoe_data.database.affixes.AffixBuilder;
 import com.robertx22.age_of_exile.aoe_data.database.affixes.ElementalAffixBuilder;
 import com.robertx22.age_of_exile.database.data.StatModifier;
 import com.robertx22.age_of_exile.database.data.gear_types.bases.BaseGearType.SlotTag;
-import com.robertx22.age_of_exile.database.data.stats.types.core_stats.AllAttributes;
-import com.robertx22.age_of_exile.database.data.stats.types.core_stats.Dexterity;
-import com.robertx22.age_of_exile.database.data.stats.types.core_stats.Intelligence;
-import com.robertx22.age_of_exile.database.data.stats.types.core_stats.Strength;
+import com.robertx22.age_of_exile.database.data.stats.types.core_stats.*;
 import com.robertx22.age_of_exile.database.data.stats.types.generated.ElementalDamageBonus;
 import com.robertx22.age_of_exile.database.data.stats.types.loot.IncreasedItemQuantity;
 import com.robertx22.age_of_exile.database.data.stats.types.loot.MagicFind;
@@ -41,30 +38,48 @@ public class JewelrySuffixes implements ISlashRegistryInit {
 
         AffixBuilder.Normal("of_the_philosopher")
             .Named("Of the Philosopher")
-            .tier(1, new StatModifier(0.2F, 0.3F, Intelligence.INSTANCE, ModType.FLAT))
-            .tier(2, new StatModifier(0.15F, 0.2F, Intelligence.INSTANCE, ModType.FLAT))
-            .tier(3, new StatModifier(0.1F, 0.15F, Intelligence.INSTANCE, ModType.FLAT))
-            .includesTags(SlotTag.jewelry_family, SlotTag.intelligence)
+            .coreStat(Intelligence.INSTANCE)
+            .includesTags(SlotTag.jewelry_family, SlotTag.cloth)
             .excludesTags(SlotTag.weapon_family)
             .Suffix()
             .Build();
 
         AffixBuilder.Normal("of_the_titan")
             .Named("Of the Titan")
-            .tier(1, new StatModifier(0.2F, 0.3F, Strength.INSTANCE, ModType.FLAT))
-            .tier(2, new StatModifier(0.15F, 0.2F, Strength.INSTANCE, ModType.FLAT))
-            .tier(3, new StatModifier(0.1F, 0.15F, Strength.INSTANCE, ModType.FLAT))
-            .includesTags(SlotTag.jewelry_family, SlotTag.strength)
+            .coreStat(Strength.INSTANCE)
+            .includesTags(SlotTag.jewelry_family, SlotTag.plate)
             .excludesTags(SlotTag.weapon_family)
             .Suffix()
             .Build();
 
         AffixBuilder.Normal("of_the_wind")
             .Named("Of the Wind")
-            .tier(1, new StatModifier(0.2F, 0.3F, Dexterity.INSTANCE, ModType.FLAT))
-            .tier(2, new StatModifier(0.15F, 0.2F, Dexterity.INSTANCE, ModType.FLAT))
-            .tier(3, new StatModifier(0.1F, 0.15F, Dexterity.INSTANCE, ModType.FLAT))
-            .includesTags(SlotTag.jewelry_family, SlotTag.dexterity)
+            .coreStat(Dexterity.INSTANCE)
+            .includesTags(SlotTag.jewelry_family, SlotTag.leather)
+            .excludesTags(SlotTag.weapon_family)
+            .Suffix()
+            .Build();
+
+        AffixBuilder.Normal("of_will")
+            .Named("Of Will")
+            .coreStat(Willpower.INSTANCE)
+            .includesTags(SlotTag.jewelry_family, SlotTag.cloth)
+            .excludesTags(SlotTag.weapon_family)
+            .Suffix()
+            .Build();
+
+        AffixBuilder.Normal("of_vitality")
+            .Named("Of Vitality")
+            .coreStat(Vitality.INSTANCE)
+            .includesTags(SlotTag.jewelry_family, SlotTag.plate)
+            .excludesTags(SlotTag.weapon_family)
+            .Suffix()
+            .Build();
+
+        AffixBuilder.Normal("of_agility")
+            .Named("Of Agility")
+            .coreStat(Agility.INSTANCE)
+            .includesTags(SlotTag.jewelry_family, SlotTag.leather)
             .excludesTags(SlotTag.weapon_family)
             .Suffix()
             .Build();