+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/affixes/adders/MagicShieldPrefixes.java
@@ -26,8 +26,8 @@ public class MagicShieldPrefixes implements ISlashRegistryInit {
 
         AffixBuilder.Normal("seraphim")
             .Named("Seraphim's")
-            .tier(1, new StatModifier(2, 3, MagicShield.getInstance(), ModType.FLAT), new StatModifier(2, 3, Health.getInstance(), ModType.FLAT))
-            .tier(2, new StatModifier(1, 2, MagicShield.getInstance(), ModType.FLAT), new StatModifier(1, 2, Health.getInstance(), ModType.FLAT))
+            .tier(1, new StatModifier(2, 3, MagicShield.getInstance(), ModType.FLAT), new StatModifier(1.5F, 2, Health.getInstance(), ModType.FLAT))
+            .tier(2, new StatModifier(1, 2, MagicShield.getInstance(), ModType.FLAT), new StatModifier(1, 1.5F, Health.getInstance(), ModType.FLAT))
             .tier(3, new StatModifier(0.5F, 1, MagicShield.getInstance(), ModType.FLAT), new StatModifier(0.5F, 1, Health.getInstance(), ModType.FLAT))
             .includesTags(SlotTag.magic_shield_stat)
             .Prefix()