+++ src/main/java/com/robertx22/age_of_exile/database/data/stats/types/defense/DodgeRating.java
@@ -55,12 +55,12 @@ public class DodgeRating extends Stat implements IStatEffects, IUsableStat, ILoc
 
     @Override
     public float getMaxMulti() {
-        return 0.8F;
+        return 0.9F;
     }
 
     @Override
     public float valueNeededToReachMaximumPercentAtLevelOne() {
-        return 200;
+        return 300;
     }
 
     private static class SingletonHolder {